nunopsilva-acme-c2ba4f3aa258/src/main/java/com/isep/acme/services/h2/RatingServiceImpl.java
package com.isep.acme.services.h2;

import com.isep.acme.services.RatingService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.isep.acme.model.Rating;
import com.isep.acme.repositories.h2.RatingRepository;

import java.util.Optional;

@Service
public class RatingServiceImpl implements RatingService {

    @Autowired
    RatingRepository repository;

    public Optional<Rating> findByRate(Double rate){
        return repository.findByRate(rate);
    }

}
